extract selectors
